@@ -1,3 +1,6 @@
+import org.gradle.api.tasks.compile.JavaCompile
+import org.gradle.api.tasks.testing.Test
+
 plugins {
     id 'java'
     id 'maven-publish'
@@ -28,13 +31,20 @@ dependencies {
     testRuntime 'ch.qos.logback:logback-classic:1.3.0-alpha5'
 
     testCompile 'org.mockito:mockito-core:3.11.2'
-    testImplementation(platform('org.junit:junit-bom:5.7.2'))
-    testImplementation('org.junit.jupiter:junit-jupiter:5.7.2')
+    testImplementation(platform('org.junit:junit-bom:5.7.0'))
+    testImplementation('org.junit.jupiter:junit-jupiter')
 }
 
 compileJava   {
-    sourceCompatibility = '1.8'
-    targetCompatibility = '1.8'
+    sourceCompatibility = 15
+    targetCompatibility = 15
+}
+
+tasks.withType(JavaCompile) {
+    options.compilerArgs += '--enable-preview'
+}
+tasks.withType(Test) {
+    jvmArgs += "--enable-preview"
 }
 
 task sourceJar(type: Jar) {
