@@ -1,4 +1,4 @@
-package net.transgressoft.commons.music;
+package net.transgressoft.commons.music.audio.operation;
 
 /**
  * @author Octavio Calleya
