@@ -1,6 +1,6 @@
 package net.transgressoft.commons.music.playlist;
 
-import net.transgressoft.commons.music.AudioItem;
+import net.transgressoft.commons.music.audio.AudioItem;
 
 import java.nio.file.Path;
 
