@@ -3,7 +3,7 @@ package net.transgressoft.commons.music.playlist;
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Objects;
 import com.google.common.collect.ImmutableSet;
-import net.transgressoft.commons.music.AudioItem;
+import net.transgressoft.commons.music.audio.AudioItem;
 
 import java.util.HashSet;
 import java.util.Optional;
@@ -18,12 +18,12 @@ import java.util.stream.Collectors;
 public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistTree<I> {
 
     private final Set<AudioPlaylist<I>> audioPlaylists;
-    private final Set<PlaylistTree<I>> subPlaylistTrees;
-    private String name;
+    private final Set<PlaylistTree<I>> includedPlaylistTrees;
+    private final String name;
 
-    protected PlaylistTreeBase(String name, Set<PlaylistTree<I>> subPlaylistTrees, Set<AudioPlaylist<I>> audioPlaylists) {
+    protected PlaylistTreeBase(String name, Set<PlaylistTree<I>> includedPlaylistTrees, Set<AudioPlaylist<I>> audioPlaylists) {
         this.name = name;
-        this.subPlaylistTrees = new HashSet<>(subPlaylistTrees);
+        this.includedPlaylistTrees = new HashSet<>(includedPlaylistTrees);
         this.audioPlaylists = new HashSet<>(audioPlaylists);
     }
 
@@ -33,38 +33,13 @@ public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistT
     }
 
     @Override
-    public void name(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean addPlaylist(AudioPlaylist<I> playlist) {
-        return audioPlaylists.add(playlist);
-    }
-
-    @Override
-    public boolean removePlaylist(AudioPlaylist<I> playlist) {
-        return audioPlaylists.remove(playlist);
-    }
-
-    @Override
-    public boolean addPlaylistTree(PlaylistTree<I> playlistTree) {
-        return subPlaylistTrees.add(playlistTree);
-    }
-
-    @Override
     public ImmutableSet<AudioPlaylist<I>> audioPlaylists() {
         return ImmutableSet.copyOf(audioPlaylists);
     }
 
     @Override
-    public ImmutableSet<PlaylistTree<I>> subPlaylistTrees() {
-        return ImmutableSet.copyOf(subPlaylistTrees);
-    }
-
-    @Override
-    public boolean removePlaylistTree(PlaylistTree<I> playlistTree) {
-        return subPlaylistTrees.remove(playlistTree);
+    public ImmutableSet<PlaylistTree<I>> includedPlaylistTrees() {
+        return ImmutableSet.copyOf(includedPlaylistTrees);
     }
 
     @Override
@@ -76,14 +51,14 @@ public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistT
         if (foundPlaylist.isPresent())
             return Optional.of(this);
 
-        Optional<PlaylistTree<I>> foundPlaylistTree = subPlaylistTrees.stream()
+        Optional<PlaylistTree<I>> foundPlaylistTree = includedPlaylistTrees.stream()
                 .filter(playlistTree -> playlistTree.name().equals(playlistName))
                 .findAny();
 
         if (foundPlaylistTree.isPresent())
             return Optional.of(this);
 
-        for (PlaylistTree<I> subPlaylistTree : subPlaylistTrees) {
+        for (PlaylistTree<I> subPlaylistTree : includedPlaylistTrees) {
             Optional<PlaylistTree<I>> result = subPlaylistTree.findParentPlaylist(playlistName);
             if (result.isPresent())
                 return result;
@@ -92,16 +67,15 @@ public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistT
         return Optional.empty();
     }
 
-    @SuppressWarnings ("unchecked")
     @Override
-    public Optional<AudioPlaylist<I>> findPlaylistByName(String playlistName) {
+    public Optional<AudioPlaylist<I>> findAudioPlaylistByName(String playlistName) {
         Optional<AudioPlaylist<I>> result = audioPlaylists.stream()
                 .filter(playlist -> playlist.name().equals(playlistName))
                 .findAny();
 
-        if (! result.isPresent()) {
-            for (PlaylistTree<I> subPlaylistTree : subPlaylistTrees) {
-                result = subPlaylistTree.findPlaylistByName(playlistName);
+        if (!result.isPresent()) {
+            for (PlaylistTree<I> subPlaylistTree : includedPlaylistTrees) {
+                result = subPlaylistTree.findAudioPlaylistByName(playlistName);
                 if (result.isPresent())
                     return result;
             }
@@ -110,15 +84,14 @@ public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistT
             return result;
     }
 
-    @SuppressWarnings ("unchecked")
     @Override
     public Optional<PlaylistTree<I>> findPlaylistTreeByName(String playlistTreeName) {
-        Optional<PlaylistTree<I>> result = subPlaylistTrees.stream()
+        Optional<PlaylistTree<I>> result = includedPlaylistTrees.stream()
                 .filter(playlist -> playlist.name().equals(playlistTreeName))
                 .findAny();
 
-        if (! result.isPresent()) {
-            for (PlaylistTree<I> subPlaylistTree : subPlaylistTrees) {
+        if (!result.isPresent()) {
+            for (PlaylistTree<I> subPlaylistTree : includedPlaylistTrees) {
                 result = subPlaylistTree.findPlaylistTreeByName(playlistTreeName);
                 if (result.isPresent())
                     return result;
@@ -129,58 +102,19 @@ public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistT
     }
 
     @Override
-    public void movePlaylist(AudioPlaylist<I> playlist, PlaylistTree<I> targetPlaylistTree) {
-        findParentPlaylist(playlist.name()).ifPresent(parentPlaylist -> {
-            parentPlaylist.removePlaylist(playlist);
-            targetPlaylistTree.addPlaylist(playlist);
-        });
-    }
-
-    @Override
-    public void movePlaylistTree(PlaylistTree<I> subPlaylistTree, PlaylistTree<I> targetPlaylistTree) {
-        findParentPlaylist(subPlaylistTree.name()).ifPresent(parentPlaylist -> {
-            parentPlaylist.removePlaylistTree(subPlaylistTree);
-            targetPlaylistTree.addPlaylistTree(subPlaylistTree);
-        });
-    }
-
-    @Override
-    public void clearPlaylistTrees() {
-        subPlaylistTrees.clear();
-    }
-
-    @Override
-    public void clearPlaylists() {
-        audioPlaylists.clear();
-    }
-
-    @Override
-    public ImmutableSet<AudioItem> audioItems() {
+    public ImmutableSet<I> audioItems() {
         Set<I> itemsFromPlaylists = audioPlaylists.stream()
                 .flatMap(playlist -> playlist.audioItems().stream())
                 .collect(Collectors.toSet());
 
-        Set<AudioItem> itemsFromSubPlaylistTrees = subPlaylistTrees.stream()
+        Set<I> itemsFromSubPlaylistTrees = includedPlaylistTrees.stream()
                 .flatMap(playlistTree -> playlistTree.audioItems().stream())
                 .collect(Collectors.toSet());
 
-        return ImmutableSet.<AudioItem>builder()
+        return ImmutableSet.<I>builder()
                 .addAll(itemsFromPlaylists)
                 .addAll(itemsFromSubPlaylistTrees)
                 .build();
-    }
-
-    @Override
-    public boolean removeAudioItems(Set<I> audioItems) {
-        boolean removed = false;
-        for (AudioPlaylist<I> audioPlaylist : audioPlaylists) {
-            removed |= audioPlaylist.removeAudioItems(audioItems);
-        }
-
-        for (PlaylistTree<I> subPlaylistTree : subPlaylistTrees) {
-            removed |= subPlaylistTree.removeAudioItems(audioItems);
-        }
-        return removed;
     }
 
     @Override
@@ -200,7 +134,7 @@ public abstract class PlaylistTreeBase<I extends AudioItem> implements PlaylistT
     public String toString() {
         return MoreObjects.toStringHelper(this)
                 .add("audioPlaylists", audioPlaylists)
-                .add("subPlaylistTrees", subPlaylistTrees)
+                .add("includedPlaylistTrees", includedPlaylistTrees)
                 .add("name", name)
                 .toString();
     }
